Nail the final remaining serialization-deserialization instability to the wall.
After about 2d of debugging I figured out that this was a result of the
serialization process not preserving the order of the Transition ArrivingArcs
and LeavingArcs arrays: the contents where the same, the order got scrambled.
This really shouldn't influence the final results, but runtimes *can* differ.

In the interests of easier performance measurement and debugging, I've now
changed the order in which arcs are serialized to keep the order in those
arrays. OTOH, the same arrays in Place instances will now get scrambled
instead. This is acceptable, as those arrays don't seem to impact state space
exploration; accordingly, I currently have no plans to expand the protocol to
keep order in all arrays.