Initial commit of io_base code.
io_base is a (relatively) simple socket-API abstraction layer intended to serve
as a base for higher-layer lola-modular networking code which has yet to be
written. It's been coded against the POSIX.1-2001 and winsock 1.0 APIs, and
empirically tested on GNU/Linux and mingw32-gcc + wine, so usage of both APIs
should at least be close to correct.
Current test code is included; makefiles and test-scripts for full automation
are To Be Written.
TBD: TCP Server and socket-dealloc-on-close code.