bug #13264:
The new dot version produces a number of carriage returns (ascii 13). This is documented in the changelog of dot as a problem since version 2.16 ("various issues with CR/LF in windows, but not in binary outputs."). 

However, CRs are now simply skipped while parsing dot files (just like LFs are), therefore fixing the bug, such that make check-png works again.